Changed Blink Behavior
Before this commit, the LED would continue to blink while the system was shutting down or rebooting. This commit causes the LED to blink one time to let the user know it's recognized the switch input. This is a bit more user friendly, but also allows for better compatibility if the user wants to run this as a systemd service.